Closes #5

+++ transfer.py
@@ -5,6 +5,7 @@ import os
 import argparse
 import uuid
 import json
+import warnings
 __author__ = 'Cox-Andrew'
 
 
@@ -19,14 +20,41 @@ def section_reader(section):
     return section_content
 
 
+# Logs in to an ICAT client using parameters specified in config, returns session ID
+#
+# :param client_name: the name of the client defined, config_name: the name of the config defined
+def client_login(client_name, config_name):
+    return client_name.login(config_name['auth'], {'username': config_name['username'], 'password': config_name['password']})
+
+
+# Returns the version of the export ICAT server as a float
+def get_version():
+    return float(json.loads(requests.get(export_config['url'] + '/icat/version').text)['version'][0:3])
+
+
+# Returns the entity transfer limit of export server
+def get_limit():
+    return json.loads(requests.get(export_config['url'] + '/icat/properties').text)['maxEntities']
+
+
 # Adds  arguments to the command line command, returns parsed arguments in object form
-def add_arguments():
+def add_arguments(toggle_limit):
     parser = argparse.ArgumentParser(description='Move an entity from one ICAT database to another')
-    parser.add_argument('query', help='Defines the ICAT entity to be exported.')
-    parser.add_argument('duplicate', choices=['throw', 'ignore', 'check', 'overwrite'],
-                        help='Defines the action to be taken if a duplicate is found. Throw: throw an exception. Ignore: go to the next row. Check: check that new data matches the old - and throw exception if it does not. Overwrite: replace old data with new.')
-    parser.add_argument('-all', action='store_true',
-                        help='All fields will be moved (default: values for modId, create Id, modDate and createdate will not be moved). This option is only available to those specified in the rootUserNames in the icat.properties file.')
+    parser.add_argument('query',
+                        help='Defines the ICAT entity to be exported.'
+                        )
+    parser.add_argument('duplicate',
+                        choices=['throw', 'ignore', 'check', 'overwrite'],
+                        help='Defines the action to be taken if a duplicate is found. Throw: throw an exception. Ignore: go to the next row. Check: check that new data matches the old - and throw exception if it does not. Overwrite: replace old data with new.'
+                        )
+    parser.add_argument('-all',
+                        action='store_true',
+                        help='All fields will be moved (default: values for modId, create Id, modDate and createdate will not be moved). This option is only available to those specified in the rootUserNames in the icat.properties file.'
+                        )
+    if toggle_limit:
+        parser.add_argument('limit',
+                            type=int,
+                            help='The maximum number of entities you want to transfer in one chunk')
     return parser.parse_args()
 
 
@@ -38,17 +66,7 @@ def attribute_assign():
         args.attributes = 'user'
 
 
-# Logs in to an ICAT client using parameters specified in config, returns session ID
-#
-# :param
-def client_login(client_name, config_name):
-    return client_name.login('simple', {'username': config_name['username'], 'password': config_name['password']})
-
-
-def get_icat_limit():
-    return json.loads(requests.get(export_config['url'] + '/icat/properties').text)['maxEntities']
-
-
+# Returns the count of entities on the ICAT, through the use of python ICAT API
 def get_entities():
     return export_client.search('SELECT count(entity) FROM ' + args.query + ' entity')[0]
 
@@ -70,13 +88,15 @@ def export_data():
 
 # Writes the exported data to specified data_file
 #
-# :param data: the exported ICAT data
+# :param data: the data to be written to disk, data_file: the filename for the file
 def write_data(data, data_file):
     with open(data_file, 'w') as f:
         f.write(data.text.encode('utf8'))
 
 
-# Streams exported data to import ICAT server, returns request operation details
+# Posts stream of data taken from specified data file
+#
+# :param data_file: the name of the file to be streamed
 def post_data(data_file):
     with open(data_file, 'rb') as stream:
         files = {
@@ -99,27 +119,45 @@ def post_data(data_file):
         )
 
 
+# Also if the operation is not successful it issues an error report
+# Returns error either 'successfully' or 'with error'
+#
+# :param return_data: the data returned by an import request
 def debug(return_data):
-    if current_pos + limit >= entities:
-        print str(entities) + ' of ' + entities + ' entities processed'
-    else:
-        print str(current_pos + limit) + ' of ' + entities + ' entities processed'
-    
-    if return_data != 204:
+    error = 'successfully'
+    if return_data.status_code != 204 and return_data.status_code != 200:
+        error = 'with error'
         return_text = json.loads(return_data.text)
-        print return_text['code'] + ' ' + return_text['message']
+        print 'Error: ' + return_text['code'] + ': ' + return_text['message']
+    return error
 
 
+# Prints out the number of entities processed as of execution
+#
+# :param error: the suffix added onto the end of the entities processed message
+def print_position(error):
+    if current_pos + limit >= entities:
+        print str(entities) + ' of ' + str(entities) + ' entities processed ' + error
+    else:
+        print str(current_pos + limit) + ' of ' + str(entities) + ' entities processed ' + error
+
+
+# Executes a series of functions to transfer the data
 def transfer_data():
     data_file = uuid_gen() + '.txt'
 
     get_return = export_data()
-    print 'Data exported'
+    if debug(get_return) == 'with error':
+        exit()
+    print 'Data exported to memory'
     write_data(get_return, data_file)
     print 'Data written to disk'
 
-    post_return = post_data(data_file)
-    debug(post_return)
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore')
+        post_return = post_data(data_file)
+
+    print_position(debug(post_return))
 
     os.remove(data_file)
 
@@ -130,19 +168,26 @@ if __name__ == '__main__':
     export_config = section_reader('export')
     import_config = section_reader('import')
 
-    args = add_arguments()
-    attribute_assign()
-
-    export_client = icat.client.Client(export_config['url'] + '/ICATService/ICAT?wsdl')
-    import_client = icat.client.Client(import_config['url'] + '/ICATService/ICAT?wsdl')
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore')
+        export_client = icat.client.Client(export_config['url'] + '/ICATService/ICAT?wsdl')
+        import_client = icat.client.Client(import_config['url'] + '/ICATService/ICAT?wsdl')
 
     export_id = client_login(export_client, export_config)
     import_id = client_login(import_client, import_config)
 
+    version = get_version()
+    if version > 4.5:
+        args = add_arguments(False)
+        limit = get_limit()
+    else:
+        args = add_arguments(True)
+        limit = args.maxEntities
+
+    attribute_assign()
+    entities = get_entities()
     data_left = True
     current_pos = 0
-    limit = get_icat_limit()
-    entities = get_entities()
 
     while data_left:
         transfer_data()
